// src/app/Medical/Models/transferapplication.ts
export class TranferapplicationMaster {
  ID: any = 0;
  CLIENT_ID = 0;
  NAME: any;
  EMPLOYEE_NAME: any;
  STATUS: any = true;
  EMPLOYEE_CODE: any;
  GRADE_PAY: any;
  OFFICE_NAME: any;
  DESIGNATION: any;
  DESIGNATION_ID: any;
  LOCATION: any;
  DDO_OF_THE_OFFICIAL: any;
  DDO_OF_THE_OFFICIAL_ID: any;
  BENEFICIARY_TYPE: string = '';
  CGHS_CARD_NO: any;
  CGHS_CARD_VALIDITY: any;
  EMP_ID: any;
  GRADE_PAY_LEVEL: any;
  GRADE_PAY_LEVEL_ID: any;
  MOBILE_NO: any;
  EMAIL_ID: any;
  ADDRESS: any;
  CURRENT_STAGE_ID: any;
  INSPECTOR_ID: any;
  NAME_OF_FAMILY_MEMBER: any;
  AGE: any;
  AD_STEP_NO: any;
  RELATIONSHIP: any;
  NEW_ADDRESS: any;
  SERVICE_TYPE: any;
  NEW_OFFICE_NAME: any;
  TRANSFER_ID: any;
  memberTransferData: any = [];
  ADVANCED_AMOUNT: any;
  ADVANCE_TAKEN_DATE: any;
  FILE_NO: any;
  TRAVEL_MODE_ID: any;
  STATION_NAME: any;
  TOTAL_AMOUNT: any;
  IS_ADVANCE_TAKEN: any;
  AMOUNT: any;
  IS_HIGHER_CLASS_ACCOMODATION: any = false;
  CERTIFIED_INFORMATION: any;
  transportationData: any = [];
  transportationDataPersonal: any = [];
  journeyData: any = [];
  FILE_URL: any;
  TRANSFER_GRANT_AMOUNT: any;
  ORDER_BILL_NO: any;
  ORDER_REASON: any;
  HEADQUARTERS_NAME: any;
  NEW_HEADQUARTERS_NAME: any;
  JOURNEY_REMARK: any;
  PERSONAL_REMARK: any;
  TRAIN_REMARK: any;
  VEHICLE_BROUGHT_SELF_PROPULATION: any = true;
  VEHICLE_BROUGHT_ROAD_MILEAGE: any;
  VEHICLE_BROUGHT_ROAD_MILEAGE_KMS: any;
  VEHICLE_BROUGHT_TOTAL: any;
  TRAIN_BROUGHT_KMS: any;
  TRAIN_BROUGHT_CHARGE: any;
  TRUCK_SHIP_CHARGE: any;
  TRAIN_BROUGHT_FOR_KMS_NO: any;
  TRAIN_BROUGHT_FOR_CHARGE_NO: any;
  ASSUMPTION_CHARGE: any;
  LPC: any;
  MANDATE_FORM: any;
  TRANSFER_ORDER_SERVICE_CERTIFICATE: any;
  TICKET_BOARDING_PASS: any;
  TRANSPORTATION_BILLS: any;
  CONDOLATION_OF_DELAY: any;
  TRANSFER_GRANT: any;

  ORDER_FORM_REMARK: any =
    'Vide F No. A. 22011/5/2022-AD-Vl, Office Order No. 212 Of 2022 dated Pg. No. of. 17/25 SI No. 20.';
  ORDER_FORM_REMARK_2: any;
  ORDER_FORM_TOTAL: any;
  ORDER_CTG: any;
  ORDER_PERSONAL_EFFECT: any;
  ORDER_AIR_FARE: any;
  ORDER_FORM_NET_TOTAL: any;
  ORDER_FORM_DATE: any;
  SIGNATURE_ID: any = 2;
  ASSUMPTION_CHARGE_STATUS: any;
  LPC_STATUS: any;
  MANDATE_FORM_STATUS: any;
  TRANSFER_ORDER_SERVICE_CERTIFICATE_STATUS: any;
  TICKET_BOARDING_PASS_STATUS: any;
  TRANSPORTATION_BILLS_STATUS: any;
  CONDOLATION_OF_DELAY_STATUS: any;
  IS_PERSONAL_CAR: any = false;
  CTG_ADMISIBLE_AMOUNT: any = 0;
  PERSONAL_EFFECT_ADMISIBLE_AMOUNT: any = 0;
  ROAD_MILEAGE_ADMISIBLE_AMOUNT: any = 0;
  JOURNEY_ADMISIBLE_AMOUNT: any = 0;
  PRIVATE_CONVIENCE_ADMISIBLE_AMOUNT: any = 0;
  CTG_REMARK: any;
  PRIVATE_CONVEYANCE_REMARK: any;
  CTG_TOTAL: any = 0;
  ROAD_MILEAGE_TOTAL: any = 0;
  JOURNEY_TOTAL: any = 0;
  PERSONAL_EFFECT_TOTAL: any = 0;
  PRIVATE_CONVEYANCE_TOTAL: any = 0;
  TRANSFER_STATUS: any = '';
  VEHICLE_CLAIMED_AMOUNT: any = 0;
  VOUCHER_NO: any = 0;
  FILE_ID: any = 0;
  ROAD_MILEAGE: any = 0;
  PER_KM: any = 0;
  WEIGHT_IN_KG: any = 0;
  RATE: any = 0;
  ROAD_MILEAGE_REMARK: any;
  INSPECTOR_NAME: any;
  TRAVEL_MODE_NAME: any;
  TRANSFER_PART_A: any;
  STEP_NO: any;
  ACTION_DATE: any = null;
  ACTION_REMARK: any = null;
  ACTION_STATUS: any = 'A';
  APP_INFO_DATE_TIME: any;
  TRANSFER_DETAILS_SUBMITTED_DATE_TIME: any;
  DOCUMENTS_SUBMITTED_DATE_TIME: any;
  DOWNLOAD_PARTA_DATE_TIME: any;
  UPLOAD_PARTA_DATE_TIME: any;
  TRANSFER_REMARK: any;
  TRANSFER_HOLD_DATE: any;
  DESTINATION_REMARK: any;
  IS_EDITABLE_DESTINATION: boolean = false;

  IS_APPLYING_FOR_ADVANCE: boolean = false;
  IS_TRANSFER_ADVANCE_CREATED: boolean = false;
  FINAL_ADVANCE_AMOUNT: any;
  DOJ_PRESENT_STATION: any;
  DOJ_PRESENT_POST: any;
  DETAILS_OF_TRANSFER_IS_IN_PUBLIC_INTEREST: any;
  PLACE_OFFICE_OF_POSTING: any;
  WHETHER_TRANSFER_IS_IN_PUBLIC_INTEREST: boolean = false;
  RESIDANCE_IS_INVOLVED: any;
  TRA_ALLO_FARE_RS: number;
  TRAVEL_DA_DAYS: number;
  TRAVEL_RS_PERDAY: number;
  CHARGES_FOR_PERSONAL_EFFECTS_AMOUNT: number;
  CHARGES_FOR_PRIVATE_CON_AMOUNT: number;
  GRANT_ONE_BASIC_PAY_AMOUNT: number;
  ADVANCE_AMOUNT_REQ: number;
  TOTAL_ENT_AMOUNT: any;
  ADVANCE_BILL_NO: any;
  ADVANCE_BILL_NO_HN: any;
  PLACE_OFFICE_OF_POSTING_HN: any;
  ADVANCE_APPROVEL_DATE: any;
  DETAILS_OF_TRANSFER_POSTING_ORDER: any;
  TRAVELLING_ALLOW_TOTAL: any;
  TRAVAL_DA_TOTAL: any;
  TRANSFER_GRANT_BASIC_PAY_TOTAL: any;
  ADVANCE_BILL_DATE: any;
  IS_TRANSFER_IN_PUBLIC_INTEREST: any;
  NEW_PROPOSED_ADDRESS: any;
  IS_EMP_FILL_ADVANCE_INFO: any = false;
  PAY_SLIP: any;
  PAY_SLIP_STATUS: any;
  SUB_INSPECTOR_ID: any;
  SALUTATION: any;
  IS_ADMIN_TRANSFER: boolean = true;
  VIDE_LETTER_NUMBER: any;
}