// src/app/Medical/Models/TransFerSignatureClass.ts
export class TransferMapData{
    ID:any;
    TRANSFER_ID: any
    TRANSFER_ORDER_SIGNATURE_1: any
    TRANSFER_ORDER_SIGNATURE_2: any
    TRANSFER_ORDER_SIGNATURE_3: any
    TRANSFER_ORDER_SIGNATURE_4: any
    
    TRANSFER_PART_B_SIGNATURE_1: any
    TRANSFER_PART_B_SIGNATURE_2: any
    TRANSFER_PART_B_SIGNATURE_3: any
    TRANSFER_CHECKLIST_SIGNATURE_1: any
    
    ADV_TRANSFER_ADV_BILL_SIGNATURE_1: any
    ADV_TRANSFER_ADV_BILL_SIGNATURE_2: any
    
    ADV_TRANSFER_MANJURI_ADESH_SIGNATURE_1: any
    ADV_TRANSFER_MANJURI_ADESH_SIGNATURE_2: any
    
    ADV_TRANSFER_DOC_SIGNATURE_1: any
    ADV_TRANSFER_DOC_SIGNATURE_2: any
    
  }